Switch away from warm + enabling xray
And other fixups.

It turns out enabling xray for a lambda busts the cache, and thus you
actually get a cold start afterwards.

Thanks lambda. Good stuff.

This moves it to a rather more involved, and hopefully more correct,
method of measuring this.

Rather than having warm and cold functions, this just creates functions
and runs them twice.

In post, I can figure out which run was cold and which was warm by
looking at the xray start-time.